Remove duplicate class, and move AsyncManualResetEvent to Sparrow
