+++ src/Sparrow/AsyncManualResetEvent.cs
@@ -1,13 +1,17 @@
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Voron.Util
+namespace Sparrow
 {
     public class AsyncManualResetEvent
     {
         private volatile TaskCompletionSource<bool> _tcs = new TaskCompletionSource<bool>();
 
-        public Task WaitAsync() { return _tcs.Task; }
+        public Task WaitAsync()
+        {
+            return _tcs.Task;
+        }
 
         public async Task<bool> WaitAsync(int timeout)
         {
